make it safer
Similar to xmipy, we keep track of the initialization state in the ModflowModel struct. Other than xmipy, we finalize and reinitialize when calling initialize on an initialized model. Similarly on finalizing a finalized model, we just don't do anything. xmipy throws errors in both cases.

Similar to xmipy, we also add the models working directory, and change directory to there in the same functions as xmipy.